refactor: Move hash exceptions to their own submodule
